+++ pysmx_stubgen/__main__.py
@@ -1,0 +1,4 @@
+from pysmx_stubgen.cli import stubgen
+
+if __name__ == '__main__':
+    stubgen()